+++ examples/gen_event_md/gen_event_md.cpp
@@ -39,7 +39,7 @@ ademco_event_t alarmEvents[] = {
 ademco_event_t excepEvents[] = {
 #define XX(name, code, zh) EVENT_##name,
 	ADEMCO_EXEPTION_EVENTS_MAP(XX)
-	ADEMCO_RESUME_EVENTS_MAP(XX)
+	ADEMCO_EXEPTION_RECOVER_EVENTS_MAP(XX)
 #undef XX
 };
 
@@ -53,8 +53,9 @@ ademco_event_t allEvents[] = {
 #define XX(name, code, zh) EVENT_##name,
 	ADEMCO_STATUS_EVENTS_MAP(XX)
 	ADEMCO_ALARM_EVENTS_MAP(XX)
+	ADEMCO_ALARM_RECOVER_EVENTS_MAP(XX)
 	ADEMCO_EXEPTION_EVENTS_MAP(XX)
-	ADEMCO_RESUME_EVENTS_MAP(XX)
+	ADEMCO_EXEPTION_RECOVER_EVENTS_MAP(XX)
 	ADEMCO_HB_EVENTS_MAP(XX)
 #undef XX
 };